requirements.txt
# Core
Pillow>=10.0.0
pystray>=0.19.0
pynput>=1.7.6
requests>=2.31.0
numpy>=1.24.0

# QR decoding (no external DLL required)
opencv-contrib-python>=4.8.0

# Screen reader + SAPI speech
accessible-output2>=0.18.0
pywin32>=306
